pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Repitis Backend - FastAPI spaced repetition learning API"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "sqlmodel>=0.0.25",
    "alembic>=1.16.5",
    "supabase>=2.21.1",
    "fsrs>=5.0.0",
    "pydantic-settings>=2.6.1",
    "python-multipart>=0.0.18",
    "psycopg2-binary>=2.9.0",
    "google-genai>=1.43.0",
    "requests>=2.32.5",
    "elevenlabs>=2.17.0",
    "uvicorn>=0.37.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
